chore: Temporarily rename administration folder
