+++ FE/app/workspaces/[workspaceId]/admin/users/page.tsx
@@ -3,7 +3,7 @@
 import { useState, useEffect } from "react";
 import { useParams } from "next/navigation";
 import { createUserColumns } from "./columns";
-import { UserTable } from "@/components/administration/UserTable";
+import { UserTable } from "@/components/administration_temp/UserTable";
 import { Button } from "@/components/ui/button";
 import { ExUpload } from "@/components/excel_import/exImportButton";
 import { Plus } from "lucide-react";
@@ -26,7 +26,6 @@ import { useGroupStore } from "@/store/groupStore";
 import { useUserStore } from "@/store/userStore";
 import { Role } from "@/apis/roleApi";
 import { Group } from "@/apis/groupApi";
-
 
 export default function UserTablePage() {
   // URL에서 workspaceId 추출
